+++ index.html
@@ -140,6 +140,6 @@
       </tbody>
     </table>
   </main>
-  <script src="index.bundle.js"></script>
+  <script src="meals.bundle.js"></script>
 </body>
 </html>
