+++ lib/foods/foods.js
@@ -1,4 +1,4 @@
-import FoodsHandler from './foods/foodsHandler'
+import FoodsHandler from './foodsHandler'
 import './foods.scss'
 
 const handler = new FoodsHandler()
