+++ lib/foods/foods.scss
@@ -1,4 +1,4 @@
-@import "./main";
+@import "../shared/main";
 
 .add-food {
   position: fixed;
