+++ lib/foods/foodsHandler.js
@@ -69,7 +69,7 @@ class FoodsHandler {
     $tr.hide()
     this.service.destroy($tr.data('id'))
       .then(() => $tr.remove())
-      .catch(() => $tr.show())
+      .catch(this.restoreData($tr))
   }
 
   handleFilterKeyup(event) {
@@ -118,6 +118,12 @@ class FoodsHandler {
       this.service.update(id, field, newValue)
         .catch(() => $td.text(oldValue))
     }
+  }
+
+  restoreData($tr) {
+    $tr.show();
+    const name = $tr.find('td.name').text()
+    alert(`${name} is part of this balanced breakfast!\nIt can't be deleted.`);
   }
 
   sortByIdDescending(list){
