+++ lib/foods/foodsService.js
@@ -1,4 +1,4 @@
-import Service from '../service'
+import Service from '../shared/service'
 
 class FoodsService extends Service {
