+++ lib/meals/meals.js
@@ -1,5 +1,5 @@
-import MealsHandler from './meals/mealsHandler'
-import './index.scss'
+import MealsHandler from './mealsHandler'
+import './meals.scss'
 
 const handler = new MealsHandler()
 handler.populate()
