+++ lib/meals/meals.scss
@@ -1,7 +1,7 @@
-// @import "./main";
+// @import "../shared/main";
 
 .negative-calories {
-  color: red;
+  color: blue;
 }
 
 .positive-calories {
