+++ lib/meals/mealsService.js
@@ -1,4 +1,4 @@
-import Service from '../service'
+import Service from '../shared/service'
 
 class MealsService extends Service {
