+++ lib/shared/_main.scss
@@ -1,4 +1,4 @@
-@import "./reset.css";
+@import "../shared/reset.css";
 @import url('https://fonts.googleapis.com/css?family=Roboto');
 
 body {
