+++ lib/shared/handler.js
@@ -1,0 +1,3 @@
+class Handler {
+  
+}
