+++ lib/shared/service.js
